refactored components structure
